Unique Set of IDs - Extensible Too
New method provides a ordered set of ID types which can be extracted to uniquely
identify each gene, the order is linked to number of genes which can be identified.
The first ID type can identify the most genes, the second identifies the most of
out of the ones which cannot yet be identified, etc.

The extra parameter can be used to pass a list of different feature types for which
to generate a list of IDs to identify these features

Signed-off-by: Sulaiman Sulaiman <SKSulai98@gmail.com>