Isolated Pseudogenes
Was mentioned in the last meeting that psuedogenes will need to be removed (yet
to do) as they will effect the results. Checked to see if they explain the large number
of genes to fail QC check, which they did! Although a few oddities still persist,
such as genes will unusual start codons

Will need to use this info to update the other scripts to remove psuedogenes

Added: Ability to find and remove pseudogenes
Updated: Refactored code to make it easier to read
Removed: Unnecessary __name__ == "__main__"
Signed-off-by: Sulaiman Sulaiman <SKSulai98@gmail.com>