Parallel Chunk Size Heuristic
Code to process the CDSs from the full database (not yet committed) was taking
upwards of +20mins. This rough heuristic aims to improve the speed of processing
for large number of jobs

For less than 10,000 jobs per process we just stick with the default heuristic
used by python, as so far it has been good enough with most jobs completing in
about 5mins or so.

This instead applies to monstriously large job sizes => A sensible number of jobs
per process felt like 10,000/4 jobs per batch, but should scale this if the job
count is significantly large: Using a log scale we can avoid quickly creating
large batches while increasing if the job count is significantly big

Signed-off-by: Sulaiman Sulaiman <ss2980@bath.ac.uk>